+++ android-sdk-ui/src/com/appboy/ui/AppboyFeedFragment.java
@@ -5,10 +5,13 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
 import android.support.v4.app.ListFragment;
+import android.support.v4.widget.SwipeRefreshLayout;
 import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.AbsListView;
 import android.widget.LinearLayout;
 import android.widget.ListView;
 import android.widget.ProgressBar;
@@ -18,13 +21,16 @@ import com.appboy.events.FeedUpdatedEvent;
 import com.appboy.events.IEventSubscriber;
 import com.appboy.models.cards.Card;
 import com.appboy.ui.adapters.AppboyListAdapter;
+import com.appboy.enums.CardCategory;
 
 import java.util.ArrayList;
+import java.util.EnumSet;
 
-public class AppboyFeedFragment extends ListFragment {
+public class AppboyFeedFragment extends ListFragment implements SwipeRefreshLayout.OnRefreshListener {
   private static final String TAG = String.format("%s.%s", Constants.APPBOY, AppboyFeedFragment.class.getName());
-  private static int NETWORK_PROBLEM_WARNING_MS = 5000;
-  private static int MAX_FEED_TTL_SECONDS = 60;
+  private static final int NETWORK_PROBLEM_WARNING_MS = 5000;
+  private static final int MAX_FEED_TTL_SECONDS = 60;
+  private static final long AUTO_HIDE_REFRESH_INDICATOR_DELAY_MS = 2500L;
 
   private final Handler mMainThreadLooper = new Handler(Looper.getMainLooper());
   // Shows the network error message. This should only be executed on the Main/UI thread.
@@ -48,6 +54,13 @@ public class AppboyFeedFragment extends ListFragment {
   private LinearLayout mEmptyFeedLayout;
   private ProgressBar mLoadingSpinner;
   private boolean mSkipCardImpressionsReset;
+  private EnumSet<CardCategory> mCategories;
+  private SwipeRefreshLayout mFeedSwipeLayout;
+
+  // This view should only be in the View.VISIBLE state when the listview is not visible. This view's
+  // purpose is to let the "network error" and "no card" states to have the swipe-to-refresh functionality
+  // when their respective views are visible.
+  private View mTransparentFullBoundsContainerView;
 
   @Override
   public void onAttach(final Activity activity) {
@@ -55,6 +68,7 @@ public class AppboyFeedFragment extends ListFragment {
     mAppboy = Appboy.getInstance(activity);
     if (mAdapter == null) {
       mAdapter = new AppboyListAdapter(activity, R.id.tag, new ArrayList<Card>());
+      mCategories = CardCategory.ALL_CATEGORIES;
     }
     setRetainInstance(true);
   }
@@ -65,6 +79,14 @@ public class AppboyFeedFragment extends ListFragment {
     mNetworkErrorLayout = (LinearLayout) view.findViewById(R.id.com_appboy_feed_network_error);
     mLoadingSpinner = (ProgressBar) view.findViewById(R.id.com_appboy_feed_loading_spinner);
     mEmptyFeedLayout = (LinearLayout) view.findViewById(R.id.com_appboy_feed_empty_feed);
+    mFeedSwipeLayout = (SwipeRefreshLayout) view.findViewById(R.id.appboy_feed_swipe_container);
+    mFeedSwipeLayout.setOnRefreshListener(this);
+    mFeedSwipeLayout.setEnabled(false);
+    mFeedSwipeLayout.setColorScheme(R.color.com_appboy_newsfeed_swipe_refresh_color_1,
+      R.color.com_appboy_newsfeed_swipe_refresh_color_2,
+      R.color.com_appboy_newsfeed_swipe_refresh_color_3,
+      R.color.com_appboy_newsfeed_swipe_refresh_color_4);
+    mTransparentFullBoundsContainerView = view.findViewById(R.id.com_appboy_feed_transparent_full_bounds_container_view);
     return view;
   }
 
@@ -84,6 +106,28 @@ public class AppboyFeedFragment extends ListFragment {
     final ListView listView = getListView();
     listView.addHeaderView(inflater.inflate(R.layout.com_appboy_feed_header, null));
     listView.addFooterView(inflater.inflate(R.layout.com_appboy_feed_footer, null));
+
+    // Enable the swipe-to-refresh view only when the user is at the head of the listview.
+    listView.setOnScrollListener(new AbsListView.OnScrollListener() {
+      @Override
+      public void onScrollStateChanged(AbsListView absListView, int scrollState) {
+
+      }
+
+      @Override
+      public void onScroll(AbsListView absListView, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
+        mFeedSwipeLayout.setEnabled(firstVisibleItem == 0);
+      }
+    });
+
+    // We need the transparent view to pass it's touch events to the swipe-to-refresh view. We
+    // do this by not consuming touch events in the transparent view.
+    mTransparentFullBoundsContainerView.setOnTouchListener(new View.OnTouchListener() {
+      @Override
+      public boolean onTouch(View view, MotionEvent motionEvent) {
+        return true;
+      }
+    });
 
     // Remove the previous subscriber before rebuilding a new one with our new activity.
     mAppboy.removeSingleSubscription(mFeedUpdatedSubscriber, FeedUpdatedEvent.class);
@@ -109,15 +153,16 @@ public class AppboyFeedFragment extends ListFragment {
 
             // If there are no cards, regardless of what happens further down, we're not going to show the list view, so
             // clear the list view and change relevant visibility now.
-            if (event.getCardCount() == 0) {
+            if (event.getCardCount(mCategories) == 0) {
               listView.setVisibility(View.GONE);
               mAdapter.clear();
             } else {
               mEmptyFeedLayout.setVisibility(View.GONE);
               mLoadingSpinner.setVisibility(View.GONE);
+              mTransparentFullBoundsContainerView.setVisibility(View.GONE);
             }
 
-            // If we our feed from offline storage, and it was old, we asynchronously request a new one from the server,
+            // If we got our feed from offline storage, and it was old, we asynchronously request a new one from the server,
             // putting up a spinner if the old feed was empty.
             if (event.isFromOfflineStorage() && (event.lastUpdatedInSecondsFromEpoch() + MAX_FEED_TTL_SECONDS) * 1000 < System.currentTimeMillis()) {
               Log.i(TAG, String.format("Feed received was older than the max time to live of %d seconds, displaying it " +
@@ -125,11 +170,12 @@ public class AppboyFeedFragment extends ListFragment {
               mAppboy.requestFeedRefresh();
               // If we don't have any cards to display, we put up the spinner while we wait for the network to return.
               // Eventually displaying an error message if it doesn't.
-              if (event.getCardCount() == 0) {
+              if (event.getCardCount(mCategories) == 0) {
                 Log.d(TAG, String.format("Old feed was empty, putting up a network spinner and registering the network error message on a delay of %dms.",
                     NETWORK_PROBLEM_WARNING_MS));
                 mEmptyFeedLayout.setVisibility(View.GONE);
                 mLoadingSpinner.setVisibility(View.VISIBLE);
+                mTransparentFullBoundsContainerView.setVisibility(View.VISIBLE);
                 mMainThreadLooper.postDelayed(mShowNetworkError, NETWORK_PROBLEM_WARNING_MS);
                 return;
               }
@@ -138,15 +184,17 @@ public class AppboyFeedFragment extends ListFragment {
             // If we get here, we know that our feed is either fresh from the cache, or came down directly from a
             // network request. Thus, an empty feed shouldn't have a network error, or a spinner, we should just
             // tell the user that the feed is empty.
-            if (event.getCardCount() == 0) {
+            if (event.getCardCount(mCategories) == 0) {
               mLoadingSpinner.setVisibility(View.GONE);
               mEmptyFeedLayout.setVisibility(View.VISIBLE);
+              mTransparentFullBoundsContainerView.setVisibility(View.VISIBLE);
             } else {
-              mAdapter.replaceFeed(event.getFeedCards());
+              mAdapter.replaceFeed(event.getFeedCards(mCategories));
               listView.setVisibility(View.VISIBLE);
             }
           }
         });
+        mFeedSwipeLayout.setRefreshing(false);
       }
     };
     mAppboy.subscribeToFeedUpdates(mFeedUpdatedSubscriber);
@@ -188,5 +236,52 @@ public class AppboyFeedFragment extends ListFragment {
     if (isVisible()) {
       mSkipCardImpressionsReset = true;
     }
+  }
+
+  public EnumSet<CardCategory> getCategories() {
+    return mCategories;
+  }
+
+  public void setCategory(CardCategory category) {
+    setCategories(EnumSet.of(category));
+  }
+
+  /**
+   * Calling this method will make AppboyFeedFragment display a list of cards where each card belongs
+   * to at least one of the given categories.
+   * When there are no cards in those categories, this method returns an empty list.
+   * When the passed in categories are null, all cards will be returned.
+   * When the passed in categories are empty EnumSet, an empty list will be returned.
+   *
+   * @param categories an EnumSet of CardCategory. Please pass in  a non-empty EnumSet of CardCategory,
+   *                   or a null. An empty EnumSet is considered invalid.
+   */
+  public void setCategories(EnumSet<CardCategory> categories) {
+    if (categories == null) {
+      Log.i(TAG, "The categories passed into setCategories are null, AppboyFeedFragment is going to display all the cards in cache.");
+      mCategories = CardCategory.ALL_CATEGORIES;
+    } else if (categories.isEmpty()) {
+      Log.w(TAG, "The categories set had no elements and have been ignored. Please pass a valid EnumSet of CardCategory.");
+      return;
+    } else if (categories.equals(mCategories)) {
+      return;
+    } else {
+      mCategories = categories;
+    }
+    if (mAppboy != null) {
+      mAppboy.requestFeedRefreshFromCache();
+    }
+  }
+
+  // Called when the user swipes down and requests a feed refresh.
+  @Override
+  public void onRefresh() {
+    mAppboy.requestFeedRefresh();
+    mMainThreadLooper.postDelayed(new Runnable() {
+      @Override
+      public void run() {
+        mFeedSwipeLayout.setRefreshing(false);
+      }
+    }, AUTO_HIDE_REFRESH_INDICATOR_DELAY_MS);
   }
 }