+++ android-sdk-ui/src/com/appboy/ui/AppboyWebViewActivity.java
@@ -3,9 +3,13 @@ package com.appboy.ui;
 import android.annotation.TargetApi;
 import android.os.Build;
 import android.os.Bundle;
+import android.view.View;
+import android.view.Window;
+import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
+
 import com.appboy.ui.activities.AppboyBaseActivity;
 
 public class AppboyWebViewActivity extends AppboyBaseActivity {
@@ -15,7 +19,12 @@ public class AppboyWebViewActivity extends AppboyBaseActivity {
   @Override
   public void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    WebView webView = new WebView(this);
+    requestWindowFeature(Window.FEATURE_PROGRESS);
+    requestWindowFeature(Window.FEATURE_INDETERMINATE_PROGRESS);
+    setContentView(R.layout.com_appboy_webview_activity);
+    setProgressBarVisibility(true);
+
+    WebView webView = (WebView) findViewById(R.id.com_appboy_webview_activity_webview);
 
     WebSettings webSettings = webView.getSettings();
     // JavaScript is enabled by default to support a larger number of web pages. If JavaScript support is not
@@ -30,6 +39,19 @@ public class AppboyWebViewActivity extends AppboyBaseActivity {
     webSettings.setUseWideViewPort(true);
     webSettings.setLoadWithOverviewMode(true);
 
+    webView.setWebChromeClient(new WebChromeClient() {
+      public void onProgressChanged(WebView view, int progress) {
+        if (progress < 100){
+          setProgressBarVisibility(true);
+        } else {
+          setProgressBarVisibility(false);
+        }
+      }
+    });
+
+    webView.getSettings().setCacheMode(WebSettings.LOAD_NO_CACHE);
+    setWebLayerTypeSafe(webView);
+
     webView.setWebViewClient(new WebViewClient());
 
     Bundle extras = getIntent().getExtras();
@@ -38,13 +60,19 @@ public class AppboyWebViewActivity extends AppboyBaseActivity {
       String url = extras.getString(URL_EXTRA);
       webView.loadUrl(url);
     }
-    setContentView(webView);
   }
 
   @TargetApi(11)
   private void setZoomSafe(WebSettings webSettings) {
     if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
       webSettings.setDisplayZoomControls(false);
+    }
+  }
+
+  @TargetApi(11)
+  private void setWebLayerTypeSafe(WebView webView) {
+    if (Build.VERSION.SDK_INT >= 11) {
+      webView.setLayerType(View.LAYER_TYPE_SOFTWARE, null);
     }
   }
 }