+++ android-sdk-ui/src/com/appboy/ui/Constants.java
@@ -1,8 +1,0 @@
-package com.appboy.ui;
-
-/**
- * Contains UI Constants used by multiple classes.
- */
-public class Constants {
-  public static final String APPBOY = "Appboy";
-}