+++ droidboy/src/com/android/vending/billing/utils/Security.java
@@ -18,9 +18,6 @@ package com.android.vending.billing.utils;
 import android.text.TextUtils;
 import android.util.Log;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
 
 import java.security.InvalidKeyException;
 import java.security.KeyFactory;